MDEV-40168: JSON-over-fulltext: move the estimator into fts0est.cc.
Pure code motion, no functional change.
